edge-gateway-modules/opcua-identity/modules/opcua_identity/requirements.txt
asyncio
azure.iot.device
